+++ private/buf/buflsp/buf_policy_yaml_lsp_test.go
@@ -88,7 +88,7 @@ func TestBufPolicyYAMLDocumentLinks(t *testing.T) {
 			absPath, err := filepath.Abs(tc.fixture)
 			require.NoError(t, err)
 
-			clientJSONConn, bufPolicyYAMLURI, _ := setupLSPServerForBufYAML(t, absPath, nil)
+			clientJSONConn, bufPolicyYAMLURI, _ := setupLSPServerForBufYAML(t, absPath, nil, nil)
 			ctx := t.Context()
 
 			var links []protocol.DocumentLink
@@ -113,7 +113,7 @@ func TestBufPolicyYAMLHoverMalformedYAML(t *testing.T) {
 	absPath, err := filepath.Abs("testdata/buf_policy_yaml/invalid/buf.policy.yaml")
 	require.NoError(t, err)
 
-	clientJSONConn, bufPolicyYAMLURI, _ := setupLSPServerForBufYAML(t, absPath, nil)
+	clientJSONConn, bufPolicyYAMLURI, _ := setupLSPServerForBufYAML(t, absPath, nil, nil)
 	ctx := t.Context()
 
 	var hover *protocol.Hover
@@ -298,7 +298,7 @@ func TestBufPolicyYAMLHover(t *testing.T) {
 	absPath, err := filepath.Abs(fixture)
 	require.NoError(t, err)
 
-	clientJSONConn, bufPolicyYAMLURI, _ := setupLSPServerForBufYAML(t, absPath, nil)
+	clientJSONConn, bufPolicyYAMLURI, _ := setupLSPServerForBufYAML(t, absPath, nil, nil)
 	ctx := t.Context()
 
 	for _, tc := range tests {