+++ private/buf/buflsp/buflsp_test.go
@@ -132,6 +132,7 @@ func setupLSPServer(
 		queryExecutor,
 		nopModuleKeyProvider{},
 		bufmodule.NopGraphProvider,
+		nopCuratedPluginVersionProvider{},
 	)
 	require.NoError(t, err)
 	t.Cleanup(func() {